+++ CHANGELOG.md
@@ -541,3 +541,6 @@
 
 * Implemented account recovery in case of deleted accounts, both generated and imported
 * Added test cases for account recovery
+* Added logs for export mnemonic
+* Added logs for export private key(for generated and imported accounts)
+* Re-encrypted the private key of imported wallets and updated vault state after change pin